requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==3.1.0
clarifai-grpc==7.11.0
emoji==2.4.0
ephem==4.1.4
googleapis-common-protos==1.59.0
grpcio==1.54.2
idna==3.4
protobuf==4.23.2
PySocks==1.7.1
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==2.0.2
